+++ modules/manager/run.go
@@ -212,33 +212,9 @@ func Run(
 		return err
 	}
 
-	var cacheOptions cache.Options
-	if cfg.CacheSyncPeriod > 0 {
-		setupLog.Info("cache sync period set", "period", cfg.CacheSyncPeriod)
-		cacheOptions.SyncPeriod = &cfg.CacheSyncPeriod
-	}
-
-	// If there are no configured watch namespaces, then we're watching ALL namespaces,
-	// and we don't have to bother individually caching any particular namespaces.
-	// This is the default behavior of the controller-runtime manager.
-	// If there are configured watch namespaces, then we're watching only those namespaces.
-	if len(cfg.WatchNamespaces) > 0 {
-		setupLog.Info("Manager set up with multiple namespaces", "namespaces", cfg.WatchNamespaces)
-		watched := make(map[string]cache.Config)
-		for _, ns := range cfg.WatchNamespaces {
-			watched[ns] = cache.Config{}
-		}
-		cacheOptions.DefaultNamespaces = watched
-	}
-
-	if cfg.ConfigMapLabelSelector != "" || cfg.SecretLabelSelector != "" {
-		cacheOptions.ByObject = map[client.Object]cache.ByObject{}
-		if err := setByObjectFor[corev1.Secret](cfg.SecretLabelSelector, cacheOptions.ByObject); err != nil {
-			return fmt.Errorf("failed to set byObject for Secrets: %w", err)
-		}
-		if err := setByObjectFor[corev1.ConfigMap](cfg.ConfigMapLabelSelector, cacheOptions.ByObject); err != nil {
-			return fmt.Errorf("failed to set byObject for ConfigMaps: %w", err)
-		}
+	cacheOptions, err := createCacheOptions(setupLog, cfg)
+	if err != nil {
+		return fmt.Errorf("failed to create cache options: %w", err)
 	}
 
 	managerOpts := ctrl.Options{